+++ okhttp/src/jsMain/kotlin/okhttp3/-JsPlatform.kt
@@ -1,0 +1,39 @@
+/*
+ * Copyright (c) 2022 Square, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package okhttp3
+
+import kotlin.coroutines.resumeWithException
+import kotlinx.coroutines.suspendCancellableCoroutine
+import okio.IOException
+
+suspend fun Call.jsExecuteAsync(): Response {
+  return suspendCancellableCoroutine { continuation ->
+    continuation.invokeOnCancellation {
+      cancel()
+    }
+    enqueue(object : Callback {
+      override fun onFailure(call: Call, e: IOException) {
+        continuation.resumeWithException(e)
+      }
+
+      override fun onResponse(call: Call, response: Response) {
+        continuation.resume(value = response, onCancellation = { call.cancel() })
+      }
+    })
+  }
+}
